# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SupportDesk API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get messages for a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message (agent or visitor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /conversations/{id}/messages/read:
    patch:
      operationId: markMessagesRead
      tags: [messages]
      summary: Mark all messages in conversation as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── Workspace ─────────────────────────────────────────────────────────────
  /workspace:
    get:
      operationId: getWorkspace
      tags: [workspace]
      summary: Get current workspace settings
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    patch:
      operationId: updateWorkspace
      tags: [workspace]
      summary: Update workspace settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspace/public/{workspaceId}:
    get:
      operationId: getPublicWorkspace
      tags: [workspace]
      summary: Get public workspace info for widget (no auth)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public workspace info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicWorkspace"

  # ── FAQs ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    # ── Auth ──────────────────────────────────────────────────────────────
...
    Workspace:
      type: object
      required: [id, name, ownerId, widgetColor, greetingMessage, botName, plan, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        ownerId:
          type: integer
        widgetColor:
          type: string
        greetingMessage:
          type: string
        botName:
          type: string
        plan:
          type: string
        stripeCustomerId:
          type: ["string", "null"]
        stripeSubscriptionId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    WorkspaceUpdate:
      type: object
      properties:
        name:
          type: string
        widgetColor:
          type: string
        greetingMessage:
          type: string
        botName:
          type: string
    PublicWorkspace:
      type: object
      required: [id, name, widgetColor, greetingMessage, botName, agentsOnline]
      properties:
        id:
          type: integer
        name:
          type: string
        widgetColor:
          type: string
        greetingMessage:
          type: string
        botName:
          type: string
        agentsOnline:
          type: boolean

    # ── Conversations ─────────────────────────────────────────────────────
...
    Message:
      type: object
      required: [id, conversationId, sender, senderType, content, seen, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        sender:
          type: string
        senderType:
          type: string
        content:
          type: string
        seen:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content, senderType]
      properties:
        content:
          type: string
        senderType:
          type: string
        sender:
          type: string

    # ── FAQs ─────────────────────────────────────────────────────────────
...